requirements/pypi.txt
# These dependencies are installed using:
#
#   pip install --upgrade
#
# This list is the default package list. All pure python packages
# for the production environment at runtime should be listed here.
# Binary (non-pure) packages may also be listed here, but you
# should see if there is a conda package that suits your needs.

six
typing
click
pathlib2

pyphen
pygments
markdown
markdown-svgbob
markdown-aafigure
markdown-blockdiag
markdown-katex
weasyprint
PyPDF2
beautifulsoup4
Jinja2

toml
PyYAML
# ujson

# https://github.com/gorakhargosh/watchdog
# From readme:
# > If you want to use an old version of Python,
# > you should stick with watchdog < 0.10.0.
watchdog<0.10.0

# https://github.com/WojciechMula/aspell-python
aspell-python-ctypes
